refactor set
