+++ internal/container/set/set.go
@@ -13,18 +13,23 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package containers
+package set
 
 // Set represents a set data structure.
 type Set[T comparable] map[T]struct{}
 
-// Add adds key into the set s.
-func (s Set[T]) Add(key T) {
-	s[key] = struct{}{}
+// New returns an initialized set.
+func New[T comparable]() Set[T] {
+	return make(Set[T])
 }
 
-// Contains returns true if the set s contains key.
-func (s Set[T]) Contains(key T) bool {
-	_, ok := s[key]
+// Add adds item into the set s.
+func (s Set[T]) Add(item T) {
+	s[item] = struct{}{}
+}
+
+// Contains returns true if the set s contains item.
+func (s Set[T]) Contains(item T) bool {
+	_, ok := s[item]
 	return ok
 }
